Split files for integration tests
