+++ tests/test_integration_0_setup_uxd_api.ts
@@ -13,7 +13,7 @@ afterEach("", () => {
     console.log("\n=====================================\n");
 });
 
-before("initialize Mango", async () => {
+before(" ======= [Suite 0 : Initialize mango (1 op)] ======= ", async () => {
     mango = await createAndInitializeMango(provider, `devnet`);
 });
 
@@ -29,7 +29,6 @@ export const depositoryWSOL = new Depository(WSOL, "SOL", SOL_DECIMALS, uxdProgr
 // Interface to the Web3 call to `UXD-Program`
 export const uxd = new UXD(provider, uxdProgram);
 
-
 // Utils Calls ----------------------------------------------------------------
 
 export async function collateralUIPriceInMangoQuote(user: NodeWallet, depository: Depository, mango: Mango): Promise<number> {
@@ -38,20 +37,12 @@ export async function collateralUIPriceInMangoQuote(user: NodeWallet, depository
 
 // Permissionned Calls --------------------------------------------------------
 
-export async function initializeControllerIfNeeded(authority: NodeWallet, controller: Controller): Promise<string> {
-    if (await provider.connection.getAccountInfo(controller.pda)) {
-        console.log("Already initialized.");
-    } else {
-        return uxd.initializeController(controller, authority, TXN_OPTS);
-    }
+export async function initializeController(authority: NodeWallet, controller: Controller): Promise<string> {
+    return uxd.initializeController(controller, authority, TXN_OPTS);
 }
 
-export async function registerMangoDepositoryIfNeeded(authority: NodeWallet, controller: Controller, depository: Depository, mango: Mango): Promise<string> {
-    if (await provider.connection.getAccountInfo(depository.mangoAccountPda)) {
-        console.log("Already registered.");
-    } else {
-        return uxd.registerMangoDepository(controller, depository, mango, authority, TXN_OPTS);
-    }
+export async function registerMangoDepository(authority: NodeWallet, controller: Controller, depository: Depository, mango: Mango): Promise<string> {
+    return uxd.registerMangoDepository(controller, depository, mango, authority, TXN_OPTS);
 }
 
 // User Facing Permissionless Calls -------------------------------------------
@@ -63,3 +54,21 @@ export function mintWithMangoDepository(user: NodeWallet, slippage: number, coll
 export function redeemFromMangoDepository(user: NodeWallet, slippage: number, amountRedeemable: number, controller: Controller, depository: Depository, mango: Mango): Promise<string> {
     return uxd.redeemFromMangoDepository(amountRedeemable, slippage, controller, depository, mango, user, TXN_OPTS);
 }
+
+// ----------------------------------------------------------------------------
+
+before("PerpMarketConfig for BTC", async () => {
+  const perpMarketConfigBTC = mango.getPerpMarketConfigFor(depositoryBTC.collateralMintSymbol);
+  const perpMarketIndexBTC = perpMarketConfigBTC.marketIndex;
+  const perpMarketBTC = await mango.group.loadPerpMarket(provider.connection, perpMarketIndexBTC, perpMarketConfigBTC.baseDecimals, perpMarketConfigBTC.quoteDecimals);
+  console.log("--- Printing the Mango BTC perp market informations ---------------- ");
+  console.log(perpMarketBTC.toPrettyString(mango.group, perpMarketConfigBTC));
+});
+
+before("PerpMarketConfig for WSOL", async () => {
+    const perpMarketConfigWSOL = mango.getPerpMarketConfigFor(depositoryWSOL.collateralMintSymbol);
+    const perpMarketIndexWSOL = perpMarketConfigWSOL.marketIndex;
+    const perpMarketWSOL = await mango.group.loadPerpMarket(provider.connection, perpMarketIndexWSOL, perpMarketConfigWSOL.baseDecimals, perpMarketConfigWSOL.quoteDecimals);
+    console.log("--- Printing the Mango BTC perp market informations ---------------- ");
+    console.log(perpMarketWSOL.toPrettyString(mango.group, perpMarketConfigWSOL));
+});
