+++ tests/test_integration_1_permissionned_2_setup_depositories.ts
@@ -1,0 +1,45 @@
+import { authority } from "./identities";
+import { expect, util } from "chai";
+import { provider } from "./provider";
+import { depositoryBTC, controllerUXD, initializeController, registerMangoDepository, depositoryWSOL, mango } from "./test_integration_0_setup_uxd_api";
+
+before(" ======= [Suite 1-1 : Test setup UXD controller (permissionned) (1 op)] ======= ", () => {
+    beforeEach("\n", async () => { });
+    afterEach("\n", async () => { });
+
+    it("Register BTC Depository to the Controller", async () => {
+        // GIVEN
+        const caller = authority;
+        const controller = controllerUXD;
+        const depository = depositoryBTC;
+
+        // WHEN
+        if (await provider.connection.getAccountInfo(depository.mangoAccountPda)) {
+            console.log("Already registered.");
+        } else {
+            const txId = await registerMangoDepository(caller, controller, depository, mango);
+            console.log(`txId : ${txId}`);
+        }
+
+        // THEN
+        depository.info();
+    });
+
+    it("Register WSOL Depository to the Controller", async () => {
+        // GIVEN
+        const caller = authority;
+        const controller = controllerUXD;
+        const depository = depositoryWSOL;
+
+        // WHEN
+        if (await provider.connection.getAccountInfo(depository.mangoAccountPda)) {
+            console.log("Already registered.");
+        } else {
+            const txId = await registerMangoDepository(caller, controller, depository, mango);
+            console.log(`txId : ${txId}`);
+        }
+
+        // THEN
+        depository.info();
+    });
+});
