+++ tests/test_integration_2_mango_depository_1_btc.ts
@@ -1,6 +1,6 @@
 import { expect } from "chai";
 import { user } from "./identities";
-import { controllerUXD, depositoryBTC, mintWithMangoDepository, redeemFromMangoDepository, mango, collateralUIPriceInMangoQuote } from "./uxdApi";
+import { controllerUXD, depositoryBTC, mintWithMangoDepository, redeemFromMangoDepository, mango, collateralUIPriceInMangoQuote } from "./test_integration_0_setup_uxd_api";
 import { printWorldInfo, printUserBalances, printDepositoryInfo, getBalance, userBTCATA, userUXDATA } from "./integration_test_utils";
 
 before("Initial world state", async () => {
@@ -8,8 +8,8 @@ before("Initial world state", async () => {
   await printUserBalances();
 });
 
-describe(" ======= [Suite one : Mint then redeem all BTC (2 op)] ======= ", () => {
-
+describe(" ======= [Suite 2-1 : Mint then redeem all BTC (2 op)] ======= ", () => {
+  beforeEach("\n", async () => { });
   afterEach("", async () => {
     await printUserBalances();
     await printDepositoryInfo(depositoryBTC, mango);
@@ -45,7 +45,7 @@ describe(" ======= [Suite one : Mint then redeem all BTC (2 op)] ======= ", () =
     expect(op1_amountBtcUsed).equals(collateralAmount * -1, "The collateral amount paid doesn't match the user wallet delta");
     expect(op1_amountUxdMinted).closeTo(maxAmountUxdMinted, maxAmountUxdMinted * (slippage), "The amount minted is out of the slippage range");
 
-    console.log(`<<<<<>>>>> ==> [Minted ${op1_amountUxdMinted} for ${op1_amountBtcUsed} BTC (prefect was ${maxAmountUxdMinted})]`);
+    console.log(`    ==> [Minted ${op1_amountUxdMinted} for ${op1_amountBtcUsed} BTC (prefect was ${maxAmountUxdMinted})]`);
   });
 
   // OP2
@@ -74,6 +74,6 @@ describe(" ======= [Suite one : Mint then redeem all BTC (2 op)] ======= ", () =
     expect(op2_amountUxdRedeemed).closeTo(maxAmountUxdRedeemed, maxAmountUxdRedeemed * (slippage), "The UXD amount redeemed is out of the slippage range");
     expect(op2_amountBtcReceived).closeTo(maxAmountBtcReceived, maxAmountBtcReceived * (slippage), "The BTC amount received is out of the slippage range");
 
-    console.log(`<<<<<>>>>> ==> [Redeemed ${op2_amountUxdRedeemed} UXD for ${op2_amountBtcReceived} BTC (perfect was ${maxAmountBtcReceived})]`);
+    console.log(`    ==> [Redeemed ${op2_amountUxdRedeemed} UXD for ${op2_amountBtcReceived} BTC (perfect was ${maxAmountBtcReceived})]`);
   });
 });
